Fix deduplication handling (#6)
* Break up `fn build_records`

* Try to fix deduplication by improving the dumb heuristic

The existing deduplication algorithm grouped packages by their repository URL,
and then used the package in that set with the shortest base name. This breaks
down in the presence of repositories like `block-modes` which might be pulled in
by both `cbc` and `ofb`, leading to an unstable result depending on hashing
order.

This change replaces the above with an attempt to reduce each of those groups of
packages to a single package based on trying to find one of the packages that
matches the repository name. If none are found, all of them are included in the
list. Additionally, this is done on the output records rather than the input
package metadata, as those records contain all the data we need.