bug: Fix handling of non-file copyright locations
If a dependency has a directory named the same as one of the copyright
locations, the license tool would try to open it as a file and exit with an
error because the tool was checking if the name existed and not if it was a
file. This change fixes the check to ensure the name exists and is a file.